renamed redux file folder to be called redux
